[#3] Remove unspecified functionality from specification.md
Problem: we don't know how whitelisting and blacklisting should
be implemented. We will get this information later. The specification
is quite big and unstable, it's hard to update it. We want to focus
on FA2 functionality now and on other token functionality later.

Solution: remove unspecified functionality for now, so that it doesn't
take our attention. Most likely will have to be heavily updated anyway.