[#3] Transfer most of the roles in one step
Problem: most of the roles can be assigned in two steps:
the owner specifies the new address for a role and then this
address should accept its role by calling `accept*` entrypoint.
For the owner role it is important because if we accidentally transfer
this role to an address we don't control we will lose full control
over the token. But for other roles it is not necessary because
the owner can reassign this role.

Solution: remove `accept*` entrypoints except for `accept_ownership`.