[#3] Prevent front-running attacks in configureMinter
Problem: if `configureMinter` is called with 100 and then
minting allowance is increased to 200,
the configured minter may insert a `mint` call between these
`configureMinter` calls and effectively mint 300 tokens.
That's not intended to work this way and is called
"front-running attack".

Solution: pass an extra argument, compare the new minting allowance
with the current one. In this case minting allowance will be increased
to 200 only if it is 100.