[#8] Add transfer operator library to stablecoin smart contract
Problem: FA2 specification requires separation of coin holders
and operators that transfer them with their corresponding transfer
policies. So, to implement proper transfer operations we need to be
able to somehow distinguish between coin holders and operators and
have some proper api to work with.

Solution: We add an additional operator storage map which stores
for every pair of owner and operator the corresponding operator
action and api for storing multiple tokens that is unused but
required by FA2 specification.