[#3] Change minting allowance to `option nat`
Problem: we ought to distinguish minters with 0 allowance and
deleted minters because the former can burn while the latter can't.
But currently this distinction is somewhat obscure in specification
of some entrypoints.

Solution: change minting allowance to `option nat` in
`get_minting_allowance` and `configure_minter`, so that `None`
represents deleted minters.