Move Haskell part into `haskell/`
Problem: the contract should be implemented in Ligo, so about
half of the repo should be in Ligo, not in Haskell.
However, we have a lot of Haskell-specific stuff in the root.
At the same time, Ligo stuff is in `ligo/`.

Solution: move everything Haskell-specific into `haskell/`.
Now `haskell/` and `ligo/` both define some `Makefile`s.
The root `Makefile` calls one of them, it has only generally useful
things. It doesn't have Haskell-specific things like `haddock`.
`make haddock` can still be called from `haskell/`.
The `LICENSE` file is duplicated. It's quite standard practice
that we use in other repos (e. g. `morley`).
The root `LICENSE` is treated specially by GitHub and the one in
`haskell/` is needed for `cabal`. And apparently symlinks do not
work.