Update FA2 specifics in the specification
Problem: FA2 was updated¹ and now contains some clarifications
regarding permissioning in `update_operators`. So some sentences
should be rephrased.

Solution: say that FA2 does not specify this logic and we stick
to the logic that is already described there.
Also clarify the uncertainty regarding the `get_token_metadata`
entrypoint.

¹ https://gitlab.com/tzip/tzip/-/merge_requests/55