[#20] Update ligo/Makefile
Problem:
1. It contains `test-ligo` target with a stub `stablecoin-test.ligo`
file, but we decided to write tests in Haskell.
2. It lacks `clean`.
3. We agreed to assume `ligo` presence in `$PATH` by default.
It is more convenient than using `docker`.
4. Currently ligo compilation target is phony, but we can make it a
normal `make` target and specify its dependencies. This way it will
be recompiled only when necessary (dependency is more recent or the
target file does not exist).

Solution:
1. Rename all leftovers of testing with LIGO. It makes `.DEFAULT_GOAL`
redundant.
2. Add a phony `clean` target.
3. Run ligo executable as just `ligo`, but make it possible to
override it using `$LIGO`.
4. Specify depdendencies of the main target, name it `stablecoin.tz`.
This way `make` will deduce when it needs to be recompiled.

As a consequence, the root `Makefile` has been updated accordingly.