Make tests depend on ligo contract
Problem: soon we will have a ligo contract and tests for it
written in Haskell. As a preparatory step we would like to setup
infrastructure for it.

Solution:
1. Update `stablecoin-test` to depend on `stablecoin.tz`.
It passes iff `stablecoin.tz` exists and can be parsed.
2. Update `haskell/Makefile` to build and symlink stablecoin.tz
before running tests.

Also remove `test-ligo` from `haskell/Makefile` because it just does
not make sense (e. g. because `stablecoin-test.ligo` is gone).