[#8] Add generic transfer hook to stablecoin contract
Problem: To follow FA2 specification requirements we need to add
some hook-related api to be able to implement transfer hooks.

Solution: We add contract hook registry environment which stores
addresses with `Set_transfer_hook` entrypoints. To validate senders
and receivers we check whether the corresponding addresses in
transfer batch have `tokens_sent` and `tokens_received` entrypoints
and skip them if not and if such is specified in our permissions
descriptor.