[Specification] Disambiguate "owner"
Problem: there is an open question about ambiguity of the word "owner"
in context of CENTRE docs and FA2. The agreed upon solution is to
say "token owner" or "contract owner".

Solution: change/fix all of "owner" occurences to be "token owner"
or "contract owner". Remove the open question as a consequence.