[Specification] Remove non-required getters
Problem: some getters that we have in the specifcation are not
required by FA2. They potentially increase contract size and gas
costs. We agreed to remove them because for off-chain access people
can use storage directly and on-chain access in unlikely to be needed
for them.

Solution:
* remove all `get_` entrypoints from the specification;
* fix incorrect name of `get_token_metadata`
(should be `token_metadata`).