[#13] Add haskell bindings for FA2 entrypoints
Problem: To start writing FA2-compatible test-suite we first need
to define some types for parameters and storage of our stablecoin
contract that represent required ones by specification.

Solution: Add types for parameter and storage of current
implementation of stablecoin smart-contract. Note that for now we
need to enforce their tree structure to match the one that ligo
generates. Also we are forced to add dummy annotations for our
storage parameters. These are known issues and for more information
about them please refer to #159 and #160 in morley repository.