[#35] Implement safelist contract interaction
Problem : Transfers can be optionally guarded by safelist contract as
described in the specification. We should implement this logic. The
contract should call the safelist contract passing the
transfer/mint/burn parameters to the safelist contract, if available in
storage. An owner callable entrypoint to set/reset the safelist should
also be implemented.

Solution : Changed transfer, mint and burn entrypoints to call the
safelist contract as specified by spec doc, and Implement the required
entrypoint to set the safelist address in storage. Added tests by adding
a dummy safelist contract.