[#35] Remove permission descriptor from storage
Problem: We have hardcoded permissions, so it is not required to hold
the permission descriptor in the storage. It also adds up to contract
origination size. So we can remove it.

Solution: Removed permission descriptor from storage and return a
hardcoded value from the query entry point for the same. Also added a
test to ensure the contract returns the expected value of the
permissions descriptor.