[#56] Refactor origination params
Problem: Right now the orignation params uses the `opBalances` field to
pass both ledger balances and operator using a map that maps token owner
address to a pair with list of operators and ledger balance. But this
means that we won't be able to add an operator for a token owner,
without adding the token owner address to the ledger.

Solution: Decouple the passing of operators to a separate field in the
`OriginationParams`. Refactor rest of the code to work with this change.