[#56] Allow zero transfer from non-existant accounts
Problem: As per FA2, the transfer operation should consider non-existant
accounts to have zero balances. But right now the contract fails if a
non existant source address is specified, even when the transfer amount
is zero. This also means that it should fail with INSUFFICIENT_BALANCE
for non-existant source accounts if the transfer is non-zero.

Solution: Fix the contract to consider current balance to be zero if the
account does not exist. If the transfer amount if greater then current
balance, throw INSUFFICIENT_BALANCE.