[#53] Cleanup minter update errors
Problem:
1. `NOT_MINTER` error is ambiguous and corresponds to two different
scenarios.
2. configure_minter errors are described a bit obscurely.

Solution: rename one of the `NOT_MINTER` errors to `ADDR_NOT_MINTER`.
It means that the address passed as an argument is not a minter,
unlike the other `NOT_MINTER` error which is about sender.
Apart from that rephrase error descriptions in the table a bit
and fix a typo.