[#63] Update tests to add checks for the errors thrown
Problem: Right now, the tests only check if the contract failed, or not
but does not check the actual error that is being thrown. But since FA2
spec and our specification required very specific errors to be thrown
for specific cases, we must check the errors thrown by the contract
matches with the specs.

Solution: Update tests with the possible error names, and use them in
tests to ensure the errors thrown mathces with what the spec prescribes.