[#63] Rename NO_ALLOWANCE_EXPECTED error
Problem: This error name was a bit ambiguous. It does not make it clear
if it meant that "an allowance was provided, when the call didn't
require it" or that "the call required an allowance, but none was
provided".

Solution: Rename this error to CURRENT_ALLOWANCE_REQUIRED, and update
specification and changelog.