[#43] SMT tests for management entrypoints
Problem: We need some SMT tests for the stablecoin contract

Solution: Implement SMT tests for the contract by implementing the
contract logic in Haskell, and applying randomly generated contract
inputs to both haskell model, and the contract running in Michleson
interpreter. After each input we check if the contract state of the
error that is thrown is the same. If they are not, the test fails.