[#45] Update morley and move operator tests to FA2 test suite
Problem: We've faced some bugs with `morley-nettest` that are
essential for resolving this issue and we need to use those
hotfixes to be able to continue working on it. Also somewhat we
our operator interaction tests are present in `Management.hs`
module which should be inside `FA2.hs`.

Solution: Update morley package and all the codebase it currently
breaks within itself.