[#77] Reuse embedded contract
Problem: there are a few places in the test code where we read the
stablecoin contract from disk, but now that we have it embedded in the
code we shouldn't need to do this anymore. This would also free us from
*having* to make sure `"test/resources/stablecoin.tz"` exists while
running the test.

Solution: Use `parseStablecoinContract` to parse the contract embedded
into the source code.