[#82] Update ligo revision
Problem: we want to check how recent ligo changes affect
the size of generated Michelson code. Also we want to keep
up with ligo updates since ligo is actively developed.

Solution: update ligo version to the latest `dev` branch
using the following command:
> niv update ligo --attribute rev=3862c0444416a2e4aba7ad116eff675cacc324d3

HEAD of `dev` was chosen because it has
https://gitlab.com/ligolang/ligo/-/merge_requests/732 that was
just recently merged and that we are interested it.