[#86] Prune zero balance and add test
Problem: Right now if an account ends up with zero balance after some
operation, it is left in the ledger. But this is not optimal, since we
interpret non-existent accounts as having zero balance anyway. So these
can be completely removed from the ledger.

Solution: In the debit_from function, check if source account is zero
after operation. If it is, remove it from the ledger. Add/Modify tests
to take new behavior into account.