[#90] Limit number of minters to 12 and store them in a map.
Problem: Currently we do not limit the number of possible minters and
thus store them in a big map. But this is not required as we won't
really have more than around 12 minters.

Solution: Limit the total number of minters to 12. Since the number of
minters is limited, we can store them in a map instead of a bigmap. So
chage the storage accordingly. Add tests to check contract won't accept
more then 12 minters.