[#91] Move metadata to external contract
Problem: As per new changes to the FA2 spec, we don't have to store the
token metadata in the contract storage. It can be moved to an external
contract's storage, and the main contract only need to store the address
of the external contract.

Solution: Implement a simple contract, with token metadata structure in
storage as defined in ligo. Change deploy process to optionally deploy
this contract on main contract deploy. Change the client command to read
the metadata to do additional indirection and fetch the metadata from
the registry contract.