[#87] Updated Haskell bindings
Problem: after adding the new TZIP-17 entrypoints to stablecoin, the
Haskell bindings no longer match the code generated by Ligo

Solution: Update the Haskell code to match the Ligo code.

Since, in order to construct a permit, the user must construct a full
contract parameter (i.e. `Lorentz.Contracts.Stablecoin.Parameter`)
rather than just the entrypoint's parameter `TransferParams`, it's now
vital that the Michelson representation of Haskell's `Parameter` and the
Michelson repr of Ligo's `closed_parameter` are an exact match.

So, we used `customGeneric` to force this.

We also added a `--default-expiry` option to the `stablecoin-client
deploy` command.