[#97] Cleanup redundant infra and tests
Problem: Currently we have some tests infra that looks for a specific
configuration of permissions before running the tests. But this is no
longer required as we are implementing specific FA2 behavior. So we can
only implement such tests and can remove the test infra for permissions
configuration check.

Solution: Remove such checks and make the `OriginationFn` to return a
TAddress instead of a Maybe, which was meant to signal an
incompatibility between the test and the contract being tested.