[#107] Updated morley to the latest version
Problem: Now that:
1. `Morley.Runtime` has been improved to avoid unnecessary typechecking,
   and
2. Our `Parameter` data type's layout is an exact match of the parameter
   type of `stablecoin.tz`

We can upgrade to the latest morley version, refactor
`parseStablecoinContract` to return a typed contract, and thus avoid
unnecessary untyping->typing steps in the SMT tests (and in other
places).

This also seems to fix the "lost signal due to full pipe: 11" error when
running the tests.

Solution: upgrade morley to the latest version, refactor
`parseStablecoinContract` and update every affected use-site.