[#102] Replace `OperatorParam` with proper ADT
Problem: The Michelson representation (that is, the balancing of `pair`
and `or`s) of our Haskell data types must match the Michelson
representation of our Ligo data types.

In the past, we used Haskell nested tuples in order to have complete
control of the layout, but this leads to messy, hard to maintain code.

Solution: Now that morley introduced `customGeneric` to control an ADT's
layout, we can replace our nested tuples with Haskell ADTs and still
have control of its layout.