[#94] Added FA1.2 SMT tests
Problem: We need a version of stablecoin that implements FA1.2 instead
of FA2, so we can compare contract size, gas & transaction costs.

Solution: Use the SMT tests from `morley-ledgers` to make sure our
implementation of FA1.2 is correct.

Note: the module `Lorentz.Contracts.Spec.FA2Interface` exists in both
the `stablecoin` and the `morley-ledgers` packages, so we need to use
the `PackageImports` extension to disambiguate.