[#102] Added compile-time check for layout of Parameter and Storage
Problem: The Michelson representation of Haskell's `Parameter` and
`Storage` must be an exact match of the Michelson repr of Ligo's
corresponding data types. They must always be kept in sync, which is a
maintenance burden.

Solution: to make this a little bit easier to maintain, we can
✨sprinkle✨ some TH and check at compile-time that the Michelson reprs
match.

Also moved `parseStablecoinContract` and `parseRegistryContract` from
`Stablecoin.Client.Contract` to `Lorentz.Contracts.Stablecoin`, because
these functions are not really related to the stablecoin-client, they
have other uses.