[#102] Removed `MonadThrow` constraint from parseStablecoinContract
Problem: Now that we're using TH to check at compile-time that the
contract can be parsed and typechecked, we should no longer need the
MonadThrow constraint on parseStablecoinContract and
parseRegistryContract.

Solution: Inside a TH splice, after proving that the contract can be
parsed+typechecked, we emit a haskell expression that reads the contract
again at runtime, except it's now safe to use `error` instead of
`throwM` because we know this can't possibly fail at runtime.