[#110] Remove is_operator entrypoint and update spec.
Problem: As per the FA2 spec, we no longer need a "Is_operator"
entrypoint. So we can remove it. Also we have to update the
specification doc to include all changes in this branch.

Solution: The `Is_operator` entrypoint is removed. Tests and
specificaton document is updated to include all the changes in the
branch.