Updated FA1.2 tests
Problem: Seeing as the purpose of FA1.2 nettest tests is to compare and
contrast the gas/transaction costs of the FA1.2 vs the FA2 version of
stablecoin, they're not currently fit for purpose. At the moment, all
they do is originate the FA1.2 contract and call its entrypoints.

Solution: Modify this test suite, so that it originates _both_ versions
of the contract, with a similar initial storage, and performs a single
`transfer` operation using similar parameters. This should ensure it's
fair to compare their gas/transaction costs.