[#113] Check metadata registry's storage layout at compile-time
Problem: In previous PRs, we added compile time checks to ensure the
layout of the Haskell's data types matches the layout of the Ligo data
types. However, we're not doing this for the metadata registry contract
yet.

Solution: Change `registryContract` to return a typed contract instead
of an untyped contract.