[#113] Refactored construction of metadata registry's storage
Problem: We have 2 functions, `mkMetadataRegistryStorage` and
`mkRegistryStorage`. The former is used only once, in the implementation
of the later. But there's no need for this indirection. 

Also, `stablecoinTokenMetadata`, should use those functions to avoid
repetition.

Solution: Merge `mkRegistryStorage` into `mkMetadataRegistryStorage`.
Replace `stablecoinTokenMetadata` so that it calls
`mkMetadataRegistryStorage` with some default values.