[#113] Use valid metadata registry contract in tests
Problem: the tests use a dummy `tz2` address as the address of the
metadata registry contract. However, the FA2 spec requires that this
address be a of a smart contract (i.e., a `KT1` address).

Solution: originate a valid metadata registry contract, and then put its
address in the stablecoin contract's storage.