CI: add 'wait' step after uploading generated contract
Problem: 'LIGO-contract' step uploads an artifact used by the next
steps, so the next steps must not start until it has finished. It wasn't
a problem when we were using a single non-concurrent agent, but when we
added an agent for scheduled pipelines, the job would fail because it
runs concurrently.

Solution: add 'wait' step to avoid running steps which use the contract
until the step which generates the contract has finished