[#127] Remove last traces of QuickCheck
Problem: we have some `Arbitrary` instances in the library code, but
they are useless because we don't use QuickCheck in our tests at all.
And we don't plan to use it. Moreover, these types should go to
morley-ledgers, but that's a task for future.
Solution: remove those instances and tasty-quickcheck dependency.