[#124] Check for too large expiry and update tests
Problem: Right now we don't check the size of expiry value in the set
expiry entrypoint. But this is dangerous as nats are unbounded in
Michelson.

Solution: Here we add a check to prevent expiry being set too large of a
value causing the contract to blow up in terms of decoding and
typechecking costs.