[#124] Enable setExpiry calls via permits
Problem: We have to limit the calls to `setExpiry` to be made only my
the permit issuer. We also have to enable call via a permit for the
`setExpire` entrypoint.

Solution: Make necessary changes to the contract code and add tests.
When called by someone other than the permit issuer, the contract now
fails with NOT_PERMIT_ISSUER error.