[#124] Updated TZIP-16 error messages
Problem: According to the TZIP-16 spec, the metadata's error messages
are meant to be shown to user.

> The interpretation is a larger data-structure, for instance, to
> provide to a wallet-user with a more understandable error message to
> act on (usually a string or bytes natural language text value).

However, our current error messages are meant to be read by developers,
and they explain _when_ those errors occur and why.

Solution: Rewrite these error messages with the end-user in mind, so as
to explain what they did wrong.