[#133] Update and polish documentation
Problem: while working on this project we were updating READMEs
somewhat gently and didn't change their structure substantially.
However, now the project is far from the stage where it was half a
year ago, we want it to be used by external people and want it to be
attractive. So I think documentation should be rethought thoroughly.
In particular, one user of the contract complained that it's not clear
how to deploy it.

Solution: I update README quite heavily with the main goal to make it
more attractive to a new person who opens the repo for the first time
and thinks whether they want to use this project.
Specifically, by looking at root README one should quickly figure out:
1. What's present in this repo.
2. How to use what's present in this repo, in particular, how to deploy
the contract. It includes an example. I believe now it should be quite
clear that stablecoin-client is recommended and it's easy to use it.
Some comments:
1. While I don't like including anything that may become outdated and
may be hard to notice, I included example usage command in README.
It may become outdated (e. g. if we add a new mandatory option), but
I think having an example is important because it shows how easy it is
to deploy Stablecoin.
2. It was a bit hard to decide what should go into the root README and
what should go into sub-READMEs.
I tried to put the most important things into the root one and some
details into sub-READMEs.