[#138] Add option to select contract metadata configuration
Problem: Right now we embed all contract (TZIP-16) metadata in contract
storage itself. But we would like to provide the user of this contract
an option to store metadata any way, as allowed by TZIP-16 spec.

Solution: Add option to the deploy command to indicate where the
contract metadata should be stored. If nothing was specified, then by
default we originate a contract for holding metadata along with the main
contract.

If the `contract-metadata-in-place` is specified then the current
behavior is followed and the metadata is embedded in the storage itself.

If the `contract-metadata-off-chain` is specified, then the user can
put the metadata in an IPFS location or host it over http/https protocol
and specify it as a string. This will be used as such, as the metadata
URI of the contract.

Net and integrational tests have been changed accomodate the changes to
the origination infra functions.