Fix SMT Test failure
Problem: An instance of SMT failure was found in a recent PR. It can be
triggerred by test arguments  '--hedgehog-replay
  "Size 18 Seed 613762895755730128 14866453392481836605"'.

Solution: It was found to be caused by a bug in the SMT Haskell model in
the burn operation in a fold call which did not use input accumilator
and insted use the starting value in scope. Fixed it by extracting the
function out and fixing the fold call.