Inline small contract functions
Problem: some small functions are created and executed as Michelson's
lambdas in the contract, because they are used in multiple places in
the LIGO source code.
However, very small functions can be more efficient, in terms of both
contract size and transaction cost, when inlined (because they have a
smaller impact that the fetching of the lambda from the stack and its
execution).

Solution: forced inlining of small functions when advantageous to the
resulting contract.
Additionally removed some unused functions as well.