Remove `all` and merge it's logic in `transfer_sender_check`
Problem: the `all` function is based on `List.fold` and is used only
in one place on a list created from a `List.map`.
This results in an (unnecessary) double traversing of the list.

Solution: remove the `all` function and merge its logic inside the one
that uses it (`transfer_sender_check`) in order to traverse the list a
single time and reduce its transaction gas cost.