[Chore] 1.5.0 release
Problem: We have provided deploy options to let the user choose where
to store the metadata, added the `total_supply` field to the contract,
inlined some functions for performance and cleaned up the docs.
For this reason, we want to make a new release.

Solution: bump Haskell package version, update changelog, update
transaction costs table in the README.