[#131] Remove token-metadata-registry from entrypoint and storage
Problem: The token metadata registry entrypoint has been removed from
TZIP-12. We have the option to keep the token metadata as annotated
bigmap in storage, but we have decided to implement it as an offchain
storage view. So we have to remove the metadata registry entrypoint as
well as the corresponding field from the storage.

Solution: Token metadata registry storage field is removed. The
entrypoint is removed as well. The tests that check for token metadata
registry field in storage is removed as well.