[#131] Add storage views
Problem: We have to add storage views as required by TZIP-12.

Solution: Implement the storage views and include them in TZIP-16
metadata. Origination procedures during tests have been changed to
remove the token metadata registry contract origination. In addition to
this, the procedure to implement the stablecoin-client commands like
`getBalance` and `getTokenMetadata` has been changed to follow the
metadata URI (currently only tezos-storage scheme is supported), and
fetch and execute the views on the storage. This is not currently
working because of inability to fetch whole storage. The
`morley-metadata` dependency has also been bumped to use the new DSL for
buidling metadata. Both the main contract metadata and FA1.2 variant has
been updated to use it.

Apart from this, some testing infra (`checkView` function) has been
moved to `Test.Common` module from the tests for permits that checks the
behavior of views. And it has been used in tests for the new storage
views.