[#131] Rename tzip16-metadata contract to metadata contract
problem: Earlier we called the contract metadata as TZIP16-metadata to
contrast with the token metadata. But now the token metadata has been
moved to tzip-16 metadata. So we only have one metadata, and we can call
it just "metadata".

solution: Remove the old metadata contract, and rename the
tzip16-metadata.ligo file as simply 'metadata.ligo'. Make the
corresponding change in code, and build configurations.