[#131] Add 'description` and `NEGATIVE_TOTAL_SUPPLY` error to Metadata
Problem: The TIZP-016 spec describe some top level fields, and we
include most of them in our Metadata, except for the 'description'
field. The 'NEGATIVE_TOTAL_SUPPLY' error is also missing from the list
of errors.

Soloution: Add description to the FA2 Metadata and include
'NEGATIVE_TOTAL_SUPPLY' in the list of errors and update tests.