[#153] Use `ligoLayout`
Problem: The generic layout of the haskell bindings are hard to maintain
because they have to be manually kept up-to-date with the layout of ligo
types.

Solution: Now that `ligoLayout` has been added to morley, we can just
use it.

We do, however, have to be careful and make sure the haskell field names
match the ligo field names, since the layout logic sorts fields
alphabetically. For this reason, we had to change `sIsPaused` to
`sPaused`.