[#153] Include ligo version in metadata
Problem: The ligo compiler is updated often, and so if we want the
compilation of a ligo contract to be reproducible we need to know which
version of the compiler it's compatible with. However, the TZIP-16
metadata for the stablecoin contract does not say which version of the
ligo compiler it's compatible with.

Solution: Inspect the nix sources.json file at compile-time, using
template haskell, and lookup which version of the ligo compiler we're
using. Then, insert it into the contract's metadata.