[#153] Added '--description' CLI option
Problem: Right now, `stablecoin-client` originates a contract with a
generic description in its metadata. But the user may sometimes want to
provide a more specific description of the contract.

Solution: add a `--description` option to `stablecoin-client`.

This option can be passed when the contract is meant to be originated
"in-place" or in a "remote contract", but it's rejected if it's passed
along with "--contract-metadata-off-chain".