[Chore] Set last Ligo version tested on a release
Problem: the ligo README currently points at a rolling release version,
which is not very convenient to get in case a new version breaks
compatibility. Moreover, there are now issues with fetching such an old
ligo release from source.

Solution: set the last tested working ligo version to 0.10 and
update the nix sources to get ligo as a static binary.