[#124] Deleted Get_counter / Get_default_expiry from FA1.2 bindings
Problem: In #135, we deleted the `get_counter` and `get_default_expiry`
views from both variants of the contract. However, we forgot to update
the lorentz bindings for the FA1.2 variant. And because we're using
`readUntypedContract` for the FA1.2 variant, this was not caught.

Solution: Update the lorentz bindings. We'll switch from
`readUntypedContract` to `embedContract` in the next commit.