[#194] Bump dependencies and update to GHC 9.0.2
Problem: stablecoin is using relatively old versions of morley & friends
and is still on GHC 8.10.7.

Solution: Update dependencies to their latest versions and to GHC 9.0.2.
Cleveland's "alias hints" have been removed, so we've switched to using
plain aliases.

We also switched from `readUntypedContract` to `embedContract` to import
the FA1.2 variant of the contract.